+++ Source/Stub.cs
@@ -1,9 +1,9 @@
 ﻿using System.Security;
 using System.Security.Permissions;
 
-///THIS IS REQUIRED CODE FOR THE STUBBED ASSEMBLY-CSHARP TO WORK.
-///This is currently required for Loader's Thunder Slam
-///
+//THIS IS REQUIRED CODE FOR THE STUBBED ASSEMBLY-CSHARP TO WORK.
+//This is currently required for Loader's Thunder Slam
+//
 #pragma warning disable CS0618
 [assembly: SecurityPermission(SecurityAction.RequestMinimum, SkipVerification = true)]
 #pragma warning restore CS0618